⚡️ Announcements and Promo Lists

+++ src/pages/admin/announcement/presenter.js
@@ -1,38 +1,67 @@
-import { makeAutoObservable } from "mobx";
+import { autorun, makeAutoObservable, runInAction } from "mobx";
 import { useAnnouncementStore } from "../../../stores/announcementStore";
 
 class AnnouncementPresenter {
+  announcement = {
+    id: new Int32Array(),
+    subject: String(),
+    message: String(),
+    event: new Int32Array(),
+  };
   announcementStore = useAnnouncementStore;
   message = "";
   subject = "";
   event_id = "";
   id = "";
+  announcements = [];
 
   constructor() {
     makeAutoObservable(this);
+    autorun(() => this.getAnnouncementList);
   }
-  get announcementList() {
-    return this.announcementStore.announcementList;
-  }
-  setFormValue = (e) => {
-    this[e.target.name] = e.target.value;
-  };
-  makeAnnouncement = async () => {
+  getAnnouncementList = async () => {
     try {
-      await this.announcementStore.makeAnnouncement({
-        message: this.message,
-        subject: this.subject,
-        event_id: this.event_id,
+      const data = await this.announcementStore.getAnnouncements();
+
+      runInAction(() => {
+        this.announcements = data.map(
+          ({ id, created_at, subject, message, events }) => ({
+            id: id,
+            subject: subject,
+            message: message,
+            event: events,
+            date: created_at,
+          })
+        );
       });
+      //console.log(data);
     } catch (error) {
       console.log(error.message);
     }
   };
-  deleteAnnouncement = async (id) => {
+  setFormValue = (e) => {
+    this.announcement[e.target.name] = e.target.value;
+  };
+  makeAnnouncement = async () => {
     try {
-      await this.announcementStore.deleteAnnouncement(id);
+      const { subject, event, message } = this.announcement;
+      let announcement = { subject, event_id: event, message };
+      console.log(announcement);
+      await this.announcementStore.makeAnnouncement(announcement);
     } catch (error) {
       console.log(error.message);
+    }
+  };
+  deleteAnnouncement = async ({ id }) => {
+    console.log(id);
+    let confirm = window.confirm("Do you want to delete this Announcement?");
+    if (confirm) {
+      try {
+        await this.announcementStore.deleteAnnouncement(id);
+        await this.getAnnouncementList();
+      } catch (error) {
+        console.log(error.message);
+      }
     }
   };
 }
